seq_added
