+++ tasks/seq/strakhov_a_char_freq_counter/include/ops_seq.hpp
