+++ tasks/seq/strakhov_a_char_freq_counter/perf_tests/main.cpp
@@ -13,20 +13,20 @@
 TEST(strakhov_a_char_freq_counter_seq, test_pipeline_run) {
   // Create data
   int expectation = 1000;
-  std::vector<char> in_string(expectation, a);
+  std::vector<char> in_string(expectation, 'a');
   std::vector<int32_t> out(1, 0);
   std::vector<char> in_target(1, 'a');
 
   // Create task_data
   auto task_data = std::make_shared<ppc::core::TaskData>();
-  if (world.rank() == 0) {
+ 
     task_data->inputs.emplace_back(reinterpret_cast<uint8_t *>(in_string.data()));
     task_data->inputs_count.emplace_back(in_string.size());
     task_data->inputs.emplace_back(reinterpret_cast<uint8_t *>(in_target.data()));
     task_data->inputs_count.emplace_back(in_target.size());
     task_data->outputs.emplace_back(reinterpret_cast<uint8_t *>(out.data()));
     task_data->outputs_count.emplace_back(out.size());
-  }
+  
   // Create Task
   auto test_task = std::make_shared<strakhov_a_char_freq_counter_seq::CharFreqCounterSeq>(task_data);
  
@@ -46,31 +46,31 @@ TEST(strakhov_a_char_freq_counter_seq, test_pipeline_run) {
   auto perf_analyzer = std::make_shared<ppc::core::Perf>(test_task);
   perf_analyzer->PipelineRun(perf_attr, perf_results);
 
-  if (world.rank() == 0) {
+ 
     ppc::core::Perf::PrintPerfStatistic(perf_results);
-  }
+  
 
-  ASSERT_EQ(expectation, out);
+  ASSERT_EQ(out, expectation);
 }
 
 TEST(strakhov_a_char_freq_counter_seq, test_task_run) {
 
   // Create data
   int expectation = 1000;
-  std::vector<char> in_string(expectation, a);
+  std::vector<char> in_string(expectation, 'a');
   std::vector<int32_t> out(1, 0);
   std::vector<char> in_target(1, 'a');
 
   // Create task_data
   auto task_data_par = std::make_shared<ppc::core::TaskData>();
-  if (world.rank() == 0) {
+ 
     task_data_par->inputs.emplace_back(reinterpret_cast<uint8_t *>(in_string.data()));
     task_data_par->inputs_count.emplace_back(in_string.size());
     task_data_par->inputs.emplace_back(reinterpret_cast<uint8_t *>(in_target.data()));
     task_data_par->inputs_count.emplace_back(in_target.size());
     task_data_par->outputs.emplace_back(reinterpret_cast<uint8_t *>(out.data()));
     task_data_par->outputs_count.emplace_back(out.size());
-  }
+  
   // Create Task
   auto test_task = std::make_shared<strakhov_a_char_freq_counter_seq::CharFreqCounterSeq>(task_data_par);
  
@@ -90,9 +90,9 @@ TEST(strakhov_a_char_freq_counter_seq, test_task_run) {
   auto perf_analyzer = std::make_shared<ppc::core::Perf>(test_task);
   perf_analyzer->TaskRun(perf_attr, perf_results);
 
-  if (world.rank() == 0) {
+ 
     ppc::core::Perf::PrintPerfStatistic(perf_results);
-  }
+  
 
-  ASSERT_EQ(expectation, out);
+  ASSERT_EQ(out, expectation);
 }
