+++ tasks/seq/strakhov_a_char_freq_counter/src/ops_seq.cpp
